feat: implement AI service interface and local smart edits (Épica K, Tasks 1-2)
- Created AIService module with comprehensive AI-assisted editing features
- Implemented local AI operations: silence detection and chapter suggestions
- Added AI models: Suggestion, AssetRef, and configuration types
- Created AIServiceTests with 50+ comprehensive tests
- Fixed compilation errors in CaptionsManagerTests, ProxyGeneratorTests, and ThumbnailCacheTests
- Integrated AIService with JobQueue, ProjectStore, and TranscriptionEngine
- Added provider-agnostic architecture for future cloud AI features

+++ EngineKit/Tests/EngineKitTests/ThumbnailCacheTests.swift
@@ -397,8 +397,8 @@ final class ThumbnailCacheTests: XCTestCase {
     }
 
     func testCachedWaveformWithDifferentResolutions() {
-        let lowResSamples = Array(repeating: 0.0, count: 100)
-        let highResSamples = Array(repeating: 0.0, count: 2000)
+        let lowResSamples = Array(repeating: 0.0 as Float, count: 100)
+        let highResSamples = Array(repeating: 0.0 as Float, count: 2000)
 
         let lowRes = ThumbnailCache.CachedWaveform(
             samples: lowResSamples,
